Made create new component form work correctly.
Now you can add components by the old way of typing text into the console and by the new way of using the form on the gui. I think the old way was faster, but the new way is good for designing a schematic on the fly (without pencil and paper).