April19,2019 Added C, L, and voltage sources to gui, and worked on circuit computations
There is a constant frequency of 1000 Hz that currently applies to batteries and voltage sources.
I am thinking about how to do mesh analysis on schemata that includes multiple voltage sources of different frequencies. The main obstacle is that I don't know how to represent voltages, impedences, and currents for multiple frequencies.